libusb-compat-1.0: avoid setting the bAlternateSetting on the wrong interface
Implement the libusb_device_handle struct to track the most recently successfully claimed interface on a given device, so that we don't risk setting the bAlternateSetting with libusb_set_interface_alt_setting() on the wrong interface, as libusb-0.1 only knows and operates on the most recently claimed one.
Also, bailout early if the interface number is negative.