nutdrv_qx: don't duplicate make targets
Since nutdrv_qx is both a serial and a USB driver, up until now, its executable and manpages appeared in two lists: serial and USB drivers, in order to be built in both cases, when ./configure'd for only one of them... but, by doing this, when ./configure'd for both, we ended up with dupes, which could potentially cause problems, e.g. when installing them (and this just happened).

For the executable, luckily, libtool calls install one file at a time, so this was not, and it still should not be, a problem.
For the manpages, we were previously saved by automake, which splits the list of files to be installed in different calls if they are too many (see automake's am__install_max var, currently set to 40), so the two nutdrv_qx entries ended up in different calls to install, but the 8b75b03 commit, by adding more manpages before nutdrv_qx's ones, pushed also the first occurrence to the second call to install, which already contained the other one, causing an error.

While other alternative, more flexible, but also less portable, ways for solving this exist (e.g. using GNU make's $(sort) on the generated lists), at least for now, prefer a more portable solution, adding new intermediate lists for serial/USB drivers and using simple automake conditionals.

Close https://github.com/networkupstools/nut/issues/594