libusb: don't auto-detach the kernel driver on FreeBSD
In FreeBSD, currently (FreeBSD from 10.2 to 11.2, at least), it is not necessary to detach the kernel driver and, further, the detaching can only be done with root privileges.
So, on FreeBSD, don't set the auto-detach flag by calling libusb_set_auto_detach_kernel_driver(), or libusb_claim_interface() will fail if the driver is not running as root.

On the other hand, even on FreeBSD, leave in place the call to libusb_kernel_driver_active() and the printing of the result, to have a useful debug info, and, in order to take into account possible future changes of direction from them, keep the explicit (non-fatal) calls to libusb_detach_kernel_driver() when claiming fails.

Also, align to nut_libusb the other USB drivers that don't rely on it to detach the kernel driver, to ease maintenance.

References:
- https://github.com/networkupstools/nut/issues/599
- https://lists.freebsd.org/pipermail/freebsd-usb/2016-March/014161.html