libusb: mark switch cases that fall through
In drivers that use libusb, where switch'ing on a LIBUSB_ERROR code, explicit the intention of falling through to the next case.
Also, to do so, use the not-pretty-but-easily-recognisable /* FALLTHRU */ format for the comment (which should also please gcc 7-ish, in case anyone is passing, directly or indirectly, -Wimplicit-fallthrough= upto 4), in case, in future, we decide to move to something else (e.g. attributes).