libusb: in open(), allow callers to call exit()
When searching for a supported device, we let callers do some things on their own to tell us if they accept or not a device, by calling the provided callback() function.
But in case that function calls exit(), we can't properly free() resources that are not visible outside the open() function, such as the list of USB devices.
So, make the list a global, and make sure to actually free() it when deinitialising.