dstate: fix dstate_delrange() socket protocol command
The documented (and used in server-side) command for deleting a range is `DELRANGE <varname> <minvalue> <maxvalue>`, not `DELRANGE <varname> "<minvalue> <maxvalue>"`, as used in dstate_delrange().
So, fix that unused (well, at least in NUT) function to use the correct format.

Also, remove an extraneous additional space in the command used in dstate_addrange().