upsd: fix the calling of netcmds items' func()
Since, as per our network protocol, it's perfectly legal to have single-word commands (and we already have a few, e.g. VER, NETVER, HELP, ...), we do not do an early check to ensure that at least one argument was provided to the command.
As such, the list of words generated by parseconf could actually end with the command itself.

That said, trying to access something that should not be there doesn't seem like a good idea, so, only pass to netcmds items' func() the list of arguments (which starts at the second item of the list of words), if there are arguments, otherwise just pass NULL.