Refresh firmware information for Eaton/HPE devices
Despite the static nature of these data, firmware can be updated by various
means, and should thus be refreshed by the driver, without needing a driver
restart. A "semi static" approach should be done in the end, as in usbhid-ups,
but is not part of the present development, though commented in the code.
For the time being, simply remove the STATIC flags to refresh the data
(snmp-ups) and allow to refresh product information (netxml-ups)

Signed-off-by: Arnaud Quette <ArnaudQuette@eaton.com>