tools/nut-usbinfo.pl: Fix handling of "bind" events
See https://lwn.net/Articles/837033/ for an in-depth explanation of how
"bind" events should be handled in udev rules.

In short:
"
The ACTION line causes the entire file to be skipped for anything other than
ADD or CHANGE events; in particular, that is what will happen with BIND
events. That will cause properties associated with those events to be
lost — and the device in question to be set up improperly (if at all).
"