main.c: write upsdrv_initinfo/updateinfo() logs to syslog
When a NUT driver is started (via upsdrvctl start), it runs
the following sequence:

1. Run as a foreground process.
2. Do some basic initialization, log to stderr.
3. If not debugging:
   - fork()
   - Run as a daemon, stderr off, syslog on.

The problem, however, is that upsdrv_initinfo() and updateinfo()
are first called before forking, as a result, the log messages is
only sent to stderr, not syslog. On a traditional init + syslog
server setup, those startup messages are lost. However, those lost
messages can be urgent - for example, it can be a warning about
a failed UPS inverter. If the server is freshly started, the admin
may miss them until much later, or keep staring at the mysterious
alarm word without finding anything relevant in syslog. Similarly,
other types of warnings, such as a rounded ondelay value, should
also be visible in syslog.

This patch checks the debugging flags and turn on syslog earlier
when it's in non-debug, non-fork mode, allowing the logs during
upsdrv_initinfo() and upsdrv_updateinfo() to be recorded in syslog.

Mailing list discussion:
https://alioth-lists.debian.net/pipermail/nut-upsdev/2021-January/007508.html

Signed-off-by: Yifeng Li <tomli@tomli.me>