Add USB device number matching
Extend the USB matching by a device number (in addition to the existing bus number)
by adding a Device field to USBDevice_[st] and ammending the RegExp by a seventh field.
Maybe some #defines for the magical constants would be appropriate.

Add DEBUG_EXACT_MATCH_DEVICE analogous to DEBUG_EXACT_MATCH_BUS, whatever its intended usage is.

Document the device field where the bus field was documented (which is not the case everywhere it's actually used).