Add Masterguard A/E driver based on Q* framework
Add a new driver for Masterguard A and E series based on the Q* framework (nutdrv_qx).

Based on vendor documentation of both the command set and USB protocol I was both asked not to disclose.

Supports "new" A700/1000/2000/3000(-19) and E40/60/100(-19) (totally untested due to lack of hardware).

Still sort of work in progress due to some open questions on NUT integration.