Update nutdrv_qx.c
Err on the safer side of changing behavior for non-hunnox device support, limit the applicability of langid_fix (per PR #638 review cautionary comments). Maybe the proposed original change did apply to other devices that support the fix-up, but that should be evaluated separately and this new constraint reversed if needed.