genericups: Add support for FTTx battery backup.
A lot of those "UPSes" use a 4-wire open-drain-like contact closure signaling (dubbed "telemetry interface") to report their status (see https://dl4jz3rbrsfum.cloudfront.net/documents/CyberPower_UM_CBN50U48A-1.pdf for example). They are sorta compatible with APC 940-0020B/C cables in terms of pin assignments, with extra signals for replace battery and battery missing and lack of shutdown signal.

This patch adds a sample device type (23) for this kind of device. It also adds support for the extra signal types (RB for replace battery and BYPASS for disconnected battery) used by these devices, as well as an option to unassign/disable an unused signal in the config file by setting it to NULL.

Note that there is no standard cable for this kind of device, so type 23 assumes RS232 control signaling and RTS as the voltage source for input pin pullups.