huawei-ups2000.c: rework logic for "load.on".
Currently the driver prevents the user from executing "load.on"
when the UPS is already on, and reports a STAT_INSTCMD_FAILED
with error logs. We need to do this if the UPS is in bypass mode,
because the UPS would otherwise enter normal mode - "load.on"
is not supposed to do that. On the other hand, executing "load.on"
when the UPS is already in "normal-mode on" has no effect, it's
still a legal command and should not be rejected arbitrarily,
which can potentially confuse downstream applications and users.

It also contains two misleading typos in the error logs: (1) it
incorrectly suggests "bypass.off" instead of "bypass.stop", and
(2) reports "failed: reason unknown." when the reason is simply
an already-on UPS.

This commit allows "load.on" to execute when it's in normal
mode and already on, and changes the log messages from errors
to warnings. Typos in the log messages are also fixed. The
"load.on" command is only rejected in bypass mode, with error
logs.

Finally, the logic for the off test is reworked to improve
readability. The original code checks whether the UPS is on by
testing the negative "not OFF" condition, which is not easy to
understand - the reader must have intimate knowledge on the code
to know its correctness. Namely, When the UPS power status is
updated, 1 of the 5 mutually exclusive flags, OL, OB, OL ECO,
BYPASS, or OFF is always set, thus the absence of the OFF flag
always reliably indicates an ON status. This commits converts
the single negative "not OFF" test to three positive tests for
"OL", "OB" and "OFF", with additional comments added on
reachable and unreachable conditions. Thus, previous uncertainty
is eliminated.

Signed-off-by: Yifeng Li <tomli@tomli.me>