Move check for LogMax encoding error to parsing of report description
At least one UPS [APC Back-UPS BX1600MI FW:294201G -302201G] is known
to have encoding errors in its USB HID report description such that a
conforming interpretation of the encoded LogicalMaximum value
indicates that it is -1 and the resulting LogicalMin..LogicalMax range
is 0..-1.

The actual report values read are correctly encoded.  Rather than log
a range error each time a value is read we can detect the encoding
error when the report description is read (once, at startup) and
adjust for the encoding error at that time.