Add a short sleep between fallback instcmds in usbhid-ups.
For some devices, missing instcmds are emulated by calling two
instcmds that do exist in sequence. For some devices, when these
commands arrive too quickly, they are not properly processed. (One,
both, or neither of the commands are acted upon.)

Adding a 1/8th-second delay (125,000 microseconds) between the two
instcmds that make up the fallback command has worked to make the
TrippLite AVR750U (new version with the 3024 productId/protocol)
behave consistently when using the fallback commands (shutdown.return
and shutdown.stayoff). This delay is also short enough that it seems
unlikely to cause a problem for any devices which don't exhibit
this issue.