Fix incorrect limits for input and config voltages in APC report descriptor
The Back-UPS XS 1400U has been observed to report input/config voltage limits
that are appropriate for the North American 120V region even though the
unit is operating in the European or other 220+V region.

This change diagnoses the the problem by checking if the logical
maximum values for UPS.Input.Voltage and UPS.Input.ConfigVoltage
are consistent with the UPS.Input.HighVoltageTransfer and if not,
increases them.

A similar problem was reported for CPS units in the EU region.

- introduces #defines for all standard usages in power system and
  battery device pages.
- moves cps-hid.c FindReport() to hidparser.c as FindObject_with_ID_Node().
- updates cps-hid.c to account for new defines/function name
- adds apc_fix_report_desc() to implement change for APC UPS units