huawei-ups2000.c: support newer UPS2000G/CH341 variants.
We received a report of the existence of a new hardware variant
from the bug tracker [1].

This UPS unit has the model number UPS2000-G-3KRTS just like
a model we've previously tested, and it has the same V2R1C1SPC50
firmware, which was tested by us as well (although not on
UPS2000-G-3KRTS but on UPS2000-A-2KTTS).

However, there are two obvious hardware differences. First, this
unit reports itself as "UPS2000G" via the protocol instead, which
is a new value not found in the datasheet or previously encountered
by us. Previously, even a "UPS2000G" model reports itself as the
"UPS2000A". Another difference is the use of WCH CH341 serial-to-USB
converter chip instead of the MaxLinear/Exar RX21V1410 chip.

    Bus 002 Device 005: ID 1a86:5523 QinHeng Electronics CH341 in serial mode, usb to serial port converter

The production label on the UPS says 2021-12-04. So it's likely a
newer hardware revision or variant. Was it redesigned to improve
USB serial compatibility? Or my personal guess, due to chip shortage?
Who knows.

Nevertheless, once the "UPS2000G" model string is added to the supported
list, it was reported that the driver correctly functions without doing
anything else. Thus, from a software perspective there's no difference.

Thanks GitHub user @Qinka for reporting it.

[1] https://github.com/networkupstools/nut/issues/1066#issuecomment-1158417669

Signed-off-by: Yifeng Li <tomli@tomli.me>