huawei-ups2000.c: rename NULL macros, close #1474.
Originally, in the struct array ups2000_cmd[], a REG_NONE macro was used
as a short-hand for a nonexistent register value. Unfortunately it was
in conflict with a Windows NT macro of the same name (previously not
found because it was only tested on Unix systems). Thus, this commit
renames it to REG_NULL. To avoid visual confusions with the function
pointer field (which also uses C's NULL for its nonexistence), this
commit also replaces it with a new macro FUNC_NULL, which is a simple
alias to C's NULL.

It's not exactly the best practice, a "UPS2000_" prefix should've been
used in principle. However, using it in this case creates exceedingly
long lines and makes the struct array definition more difficult to read,
which defeats the very purpose of defining these macros to begin with.
Considering that it's only an one-off use, it's unlikely to cause further
problems.

Signed-off-by: Yifeng Li <tomli@tomli.me>