huawei-ups2000: fix broken timeout recovery behavior in Issue #1846.
Occasionally, the UPS takes longer than 1 second to respond
and finish transmitting a message. This causes a timeout in
modbus_read_registers(), and the driver is supposed to retry
the request and recover from the failure. Unfortunately, when
the driver retries, instead of getting the result of the current
request, the leftover bytes from the previous read is received
from the serial buffer, creating invalid messages. It causes
all following reads to also fail, often with an "Invalid CRC"
error, and the driver can never recovery from the failure.

This commit fixes the bug by flushing the buffer using function
modbus_flush() before invoking modbus_read_registers(). The
libmodbus timeout value is also bumped to 2 seconds to avoid
unnecessary timeouts and subsequent retries. Finally, the
version number is also bumped to reflect the existence of this
fix.

Signed-off-by: Yifeng Li <tomli@tomli.me>