tripplite_usb.c add device matching based on ups.id
Implemented a function: match_ups_id to read the upsid from config file, read the ups.id from the device and compare them such that the driver can match based on the ups.id (unit id) which is user configureable using upsrw. We need to add this to driver documentation.