@jimklimov my apologies i royally screwed up the pull request with a rebase. I just started a new to have a clean slate scnario. also added the NUT_UNUSED_VARIABLE  per your recommendation.
tripplite_usb.c Added UPS ID (upsid) Support: The code now includes support for matching and uniquely identifying UPS devices using the UPS ID (upsid). The upsid configuration option has been added, which accepts a regular expression to match the UPS ID string.
This was done by adding a new function called match_by_unitid, that reads the upsid from the config file, then reads the upsid from the device, and checks if they match. This function is now passed to the comm_driver matcher on first connect and reconnects.

Added the upsid to the upsdrv_makevartable function so it is read and stored from the config file.

tripplite_usb.txt Updated Documentation: The driver documentation was updated to reflect the new functionality. The "EXTRA ARGUMENTS" section now includes information about the upsid configuration option, and how to set it.

thank you to @jimklimov for pointing me in the right direction and hand holding.

Signed-off-by: sapireli <114778791+sapireli@users.noreply.github.com>