apc_modbus: Use snprintf instead of strncpy
`strncpy` does not necessarily zero terminate the buffer so we replace
it with `snprintf` and add error checking.

This also fixes an off-by-one in the `snprintf` error checking of the
`_apc_modbus_double_to_nut` and `_apc_modbus_power_to_nut` functions and
adds error checking to every `snprintf` call.

Signed-off-by: Axel Gembe <axel@gembe.net>