Add usb_config_index to usb_communication_subdriver_s, default to 0.
The tactical goal of this change is to change
                ret = libusb_get_config_descriptor(device,
                        (uint8_t)usb_subdriver.hid_rep_index,
                        &conf_desc);
to
                ret = libusb_get_config_descriptor(device,
                        (uint8_t)usb_subdriver.usb_config_index,
                        &conf_desc);

Before this change, libusb1.c did libusb_get_config_descriptor()
with a config index equal to the interface number. For composite
devices using an interface index > 0, this is usally the wrong
choice. Concretely, I'm using an Arduino for a DIY UPS project and
these are composite devices with multiple interfaces under the
first (and only) config descriptor.

In the USB descriptor heirarchy, deivce descriptors have config
descriptors which have interface descriptors. Also, nearly all
USB devices have a single configuration (index 0).

In order to do this, I added a new member alongside the existing
hid_rep_idex and hid_desc_index. I chose to do this instead of
using the add_var method because this member is used in places
in very similar ways to how hid_rep_index and hid_desc_index is
used. This new member defaults to 0 which covers the majority of
USB devices. Any future subdriver is able to use this if a device
requires it. For existing subdrives, we'll just use an index of 0.

I also changed some debug logging to print out the config index
where the code was already printing the interface index.

Signed-off-by: Kelly Byrd <kbyrd@memcpy.com>