common/nutstream.cpp, include/nutstream.hpp: NutFile class: refactor constructor with access_t mode to not use tmpnam() [#2295]
The method for NutFile::tmpName() was private and only used in this
constructor, and relied on the unsafe tmpnam() method to openx()
the resulting file with a given mode.

With this change, the file is created via tmpfile() and freopen()'ed
with the desired mode.

Downside: name inherently unknown.

Upside: auto-deletion of the temporary file.

Closes: #2295

Signed-off-by: Jim Klimov <jimklimov+nut@gmail.com>