common/nutstream.cpp: avoid use of hard-coded temporary path; fix WIN32 use-case [#2293]
In WIN32 builds, ::tmpfile() tries to write to C:\ root dir and usually has no permissions to do so.
Here we actively probe common envvars for temporary directories (or also use a WIN32 syscall for the effect) to try and get a realistic TMPDIR.

TODO: Consider NUT statepath/(alt)pidpath... maybe not available in tests though. Probe if the chosen dir actually exists?

Signed-off-by: Jim Klimov <jimklimov+nut@gmail.com>