bicker_ser: expose mappable parameters
Every Bicker parameter that has a correspondence in nut-names.txt has
been mapped. If the parameter is disabled on the device, no NUT variable
is created. If the NUT variable is set to an empty string, the parameter
is reset and disabled on the device and that NUT variable is removed.

Signed-off-by: Nicola Fontana <ntd@entidi.it>