bicker_ser: allow NULL dst in parameter handling
The comments say to use `NULL` for discarding the results but
bicker_receive_parameter() was accessing `dst->id`, effectively
preventing it.

Fix that bug and improve the code: now bicker_receive_parameter()
validates the parameter id (so it can be used freely) and it always logs
debugging info about the parameter.

Signed-off-by: Nicola Fontana <ntd@entidi.it>