server/netget.c: sanity-check NUMBER values that they actually resolve as float or long int [#266]
If we assume a value without flags is by default a NUMBER but is not,
re-default it to STRING. If it is also flagged explicitly as a NUMBER
but is not, yell a warning for upstream to fix the code. Eventually
this should be a reason to report protocol error, but not now.

Signed-off-by: Jim Klimov <jimklimov+nut@gmail.com>