libusb0/1: Simplify range checks in nut API nut_libusb_get_string()
The string index is a number between 1 and 255, simply because it is the
lower byte of wValue in the GET_DESCRIPTOR request. String index 0 is
used for retrieving the langid array, so it is not valid for this
function.

There is no reason to impose a limit to the length of the buffer that
the user is lending us, certainly not above the range of the involved
types, so just a minimal sanity check is enough. Since in the minimal
case we would return an empty zero-terminated string, it must at least
have room for this.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>