drivers/libusb1.c: fix wrong return value of nut_libusb_get_interrupt()
cf github #2401.

When global var 'interrupt_size' is not set, a transfer size of SMALLBUF (512) is requested when calling libusb_interrupt_transfer(). Our function nut_libusb_get_interrupt() should return the actual transferred size. Otherwise, the caller will attempt to parse any uninitialized data in the buffer after the response.

In libusb0.c this was done correctly by using the return value of usb_interrupt_read(). With libusb1, the transferred size is written to tmpbufsize, and should therefore be used as the return val.

Signed-off-by: fenugrec <fenugrec@users.sourceforge.net>