Makefile.am: redefine SUBDIRS list to handle docs after code builds
On the one hand, code builds are well-parallelized, easy to avoid extra
work when re-making, and are more useful in typical development cycles.
The ChangeLog HTML/PDF generation is a time-consuming hog when enabled,
and with automake's "all-recursive" implementation handling SUBDIRS only
sequentially, it blocks the useful part of build for a long time when we
e.g. `git commit` the code changes during development.

Signed-off-by: Jim Klimov <jimklimov+nut@gmail.com>