drivers/nutdrv_hashx: Add new driver for PowerManager+ supported devices
Add an initial driver for the "#*" protocol support that is used by many
UPS out there that rely on the proprietary software suites PowerMaster+,
PowerMaster, PowerGuide and so on...

The naming of the driver has been chosen since all the UPS replies start
with an "#", while there's not a definitive common pattern for the queries
or commands.

Apparently the same protocol is used by various UPS devices, both online
and line-interactive (given that many producers advertise to use the
said programs), and only in serial mode (AFAIK). Although this driver has
solely been tested with the Atlantis Land S1200 line-interactive UPS.

So there may be missing features (such as the bypass mode) or lack of
information. However most of the decoding has been done and the driver
is enough to update the UPS status for monitoring and shutdown purposes.

As for now, only the status updates features have been implemented.

Signed-off-by: Marco Trevisan (Treviño) <mail@3v1n0.net>

Closes: #2769
Signed-off-by: Marco Trevisan (Treviño) <mail@3v1n0.net>