tools/nut-scanner/Makefile.am: ensure we only call NUT_SCANNER_DEPS and it has work to do only once [follows up from PR #2825]
In practice we do call it as many times as there are listed
dependency files, but it only generates them on the first run
and quicky bails out later on. This should help avoid writing
into same files in parallel, ending up with garbled or incomplete
files when we want to use them.

Signed-off-by: Jim Klimov <jimklimov+nut@gmail.com>