nutdrv_qx_masterguard: Fix masterguard_add_slaveaddr return value
The masterguard_add_slaveaddr function, defined as
a preprocess_command handler, must return the length of the command
string it has processed. It was incorrectly returning 0.

This caused the main qx_process function to see a 0-byte command
length, resulting in a send: timeout (0) debug log and a failure
to send the command to the UPS.

This commit fixes the function to return the correct command length
(which was already stored in the local variable l), resolving the bug
for all commands that use the ,XX slave address placeholder.

Tested on Masterguard A1000 and A2000 units.

Signed-off-by: Atanas Vladimirov <vladimirov.atanas@gmail.com>